remover produto feito, funcionario e clientes não funcionando
remover funcionário e clientes não estão funcionando pois as tabelas filhas não estão em seus objetos

+++ src/app/material/pages/consultar-material/consultar-material.component.ts
@@ -45,18 +45,18 @@ export class ConsultarMaterialComponent implements OnInit {
 
    public remover(material: MaterialDTO){
     this.usuario = JSON.parse(sessionStorage.getItem("usuarioSessao"));
-    console.log(material);
-    this.service.deleteMaterial(material,this.usuario).subscribe(
-      sucesso => {
-        console.log(sucesso);
-        console.log(this.materiais);
-        this.buscarMateriais();
-      },
-      erro => {
-        console.log(erro)
-        alert(erro);
-      }
-    );
+    if(confirm(`Deseja realmente excluir o material "${material.nome}"?`)){
+      this.service.deleteMaterial(material,this.usuario).subscribe(
+        sucesso => {
+          console.log(sucesso);
+          console.log(this.materiais);
+          this.buscarMateriais();
+        },
+        erro => {
+          console.log(erro)
+        }
+      );
+    }
   }
